Add replayable move history
This is a very early prototype. We are aware of the following scenario:

  Player A is currently replaying past moves while it's Player B's turn.
  If Player B makes a move, then Player A's board will instantly jump
  forward to the new move.

While this behavior isn't entirely useless, it can be bad UX. The plan
is to not have Player A's board instantly jump, while ensuring that
Player A is properly notified that Player B has made a move.